% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_bootstrap.R
\name{balanced_bootstrap}
\alias{balanced_bootstrap}
\alias{balanced_bootstrap.data.frame}
\alias{balanced_bootstrap.grouped_df}
\title{Generate balanced bootstrap replicates}
\usage{
balanced_bootstrap(data, ...)

\method{balanced_bootstrap}{data.frame}(data, k = 1L, ...)

\method{balanced_bootstrap}{grouped_df}(data, k = 1L, stratify = FALSE, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{k}{Number of replicates}

\item{stratify}{If \code{TRUE}, resample within groups, otherwise,
resample by groups instead of rows.}
}
\value{
A data frame with \code{k} rows and columns:
\describe{
\item{\code{sample}}{A list of \code{\link{resample}} objects with
with the bootstrap replicates.}
\item{\code{.id}}{An integer vector with replicate ids}
}
}
\description{
A balanced bootstrap generates \code{k} replicates in which each element
appears \code{k} times.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Balanced bootstrap of rows in a data frame.

\item \code{grouped_df}: Balanced bootstrap either by group, or
within groups (if \code{stratify = TRUE}).
}}

\references{
\itemize{
\item{Gleason, John R. 1988. "Algorithms for Balanced Bootstrap Simulations." \emph{The American Statistician}
\href{https://dx.doi.org/10.2307/2685134}{doi:10.2307/2685134}.}
\item{Davison, A.C., Hinkley, D.V., and Schechtman, E. 1986.
"Efficient Bootstrap Simulation." \emph{Biometrika}.}
}
}
