% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_bias}
\alias{boot_bias}
\title{Bootstrap bias estimate}
\usage{
boot_bias(boot)
}
\arguments{
\item{boot}{An object returned by \code{\link{boot_stat}}.}
}
\value{
Numeric scalar, bootstrap estimate of the bias \eqn{E^*[T^*]-T}.
}
\description{
Bootstrap bias estimate
}
