% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_se}
\alias{boot_se}
\title{Bootstrap standard error}
\usage{
boot_se(boot)
}
\arguments{
\item{boot}{An object returned by \code{\link{boot_stat}}.}
}
\value{
Numeric scalar, bootstrap estimate of the standard error of \eqn{T}.
}
\description{
Bootstrap standard error
}
