% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap_n}
\title{Generate bootstrap replicates}
\usage{
bootstrap(data, R = 1L, n = NULL, extractor = NULL, ...)

bootstrap_n(n, R = 1L, weights = NULL, bayes = FALSE, m = NULL,
  replace = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{R}{Number of replicates to sample}

\item{...}{Passed to methods}

\item{weights}{Observation level weights. This must be the name of a column
in \code{data}.}

\item{bayes}{If \code{TRUE}, a Bayesian bootstrap is used.}

\item{m}{Number of observations in the replicates.}

\item{stratify}{Resample within groups (stratified bootstrap)}

\item{groups}{Resample groups (clustered bootstrap)}

\item{weight_groups}{If \code{TRUE} and \code{groups = TRUE},
then sample groups using group-level weights.
If \code{weights} is \code{NULL}, then weight groups by the number
of observations in each. If \code{weights}, is not-\code{NULL}, then
use the sum of the observation level weights in each group as the
group-level weight.}

\item{weight_within}{If \code{TRUE} and \code{stratify = TRUE}, then
use observation level weights to sample observations within each group.}
}
\description{
Generate bootstrap replicates. In addition to the ordinary bootstrap,
it supports weighted bootstrap (\code{weights}),
Bayesian bootstrap (\code{bayes}), cluster bootstrap (if \code{data} is a
grouped data frame and \code{groups = TRUE}), and stratified bootstrap
(if \code{data} is grouped data frame and \code{stratify = TRUE}).
}
\examples{
# example from modelr::bootstrap
library("purrr")
library("dplyr")
boot <- bootstrap(mtcars, 100)

models <- map(boot$sample, ~ lm(mpg ~ wt, data = .))
tidied <- map_df(models, broom::tidy, .id = "id")

hist(subset(tidied, term == "wt")$estimate)
hist(subset(tidied, term == "(Intercept)")$estimate)
}
\references{
\itemize{
\item{Angelo Canty and Brian Ripley (2016). boot: Bootstrap R (S-Plus) Functions. R package version 1.3-18.}
\item{Davison, A. C. & Hinkley, D. V. (1997) Bootstrap Methods and Their Applications. Cambridge University Press, Cambridge. ISBN 0-521-57391-2}
}
}
\seealso{
The \pkg{boot} function \code{\link[boot]{boot}} which is the
  canonical R bootstrap implementation. This function produces slightly different
  output and has more options than the \pkg{modelr} function
  \code{\link[modelr]{bootstrap}}.
}
