% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_df.R
\name{crossv_df}
\alias{crossv_df}
\alias{crossv_df.data.frame}
\alias{crossv_df.grouped_df}
\title{Generate cross-validation test-training sets manually}
\usage{
crossv_df(data, ...)

\method{crossv_df}{data.frame}(data, train = NULL, test = NULL, ...)

\method{crossv_df}{grouped_df}(data, train = NULL, test = NULL, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{train}{A list of integer vectors or an integer vector, with the indices of the training set.
If \code{train} is \code{NULL}, then the training set is the complement of
the test set.}

\item{test}{A list of integer vectors or an integer vector, with the indices of the test set.
If \code{test} is \code{NULL}, then the test set is the complement of
the training set.}
}
\description{
Generate cross-validation test-training sets from manually specified
indices of observations in the test and and training sets.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Partition a data frame into test and training sets by row.

\item \code{grouped_df}: Partition a grouped data frame into test and training
sets by group.
}}

\examples{
library("dplyr")

# test-training pairs can be specified by test sets
crossv_df(mtcars, train = list(1:5, 6:10))
# or by training sets
crossv_df(mtcars, train = list(1:20, 10:32))
# or by both (and they don't need to partition the observations)
crossv_df(mtcars,
          train = list(1:20, 10:32),
          test = list(21:32, 1:9))

# with grouped data frames, the indices refer to groups
crossv_df(group_by(mtcars, mpg), train = list(1L))$train
}
