% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold.R
\name{crossv_kfold}
\alias{crossv_kfold}
\alias{crossv_kfold.data.frame}
\alias{crossv_kfold.grouped_df}
\title{Generate cross-validated k-fold test-training pairs}
\usage{
crossv_kfold(x, k, ...)

\method{crossv_kfold}{data.frame}(x, k = 5L, shuffle = TRUE, ...)

\method{crossv_kfold}{grouped_df}(x, k = 5L, shuffle = TRUE,
  stratify = FALSE, ...)
}
\arguments{
\item{k}{The number of folds}

\item{...}{Passed to methods}

\item{shuffle}{If \code{TRUE}, randomly assign observations to folds.
Otherwise, observations are sequentially assigned to folds.}

\item{stratify}{If \code{TRUE}, within each group observations are split
into folds, and those folds combined. If \code{FALSE}, groups are assigned
into folds.}

\item{data}{A data frame}
}
\value{
A data frame with \code{k} rows and the following columns:
\describe{
\item{train}{A list of \code{\link{resample}} objects. Training sets.}
\item{test}{A list of \code{\link{resample}} objects. Test sets.}
\item{.id}{Character vector. ID of the fold.}
}
}
\description{
Generate cross-validated k-fold test-training pairs
}
\examples{
cv <- crossv_kfold(mtcars, 5)
cv

library("purrr")
models <- map(cv$train, ~ lm(mpg ~ wt, data = .))
errs <- map2_dbl(models, cv$test, modelr::rmse)
hist(errs)

# stratified cross-validation
library("dplyr")
cv2 <- crossv_kfold(group_by(mtcars, am), 5, stratify = TRUE)
cv2
models2 <- map(cv2$train, ~ lm(mpg ~ wt, data = .))
errs2 <- map2_dbl(models, cv2$test, modelr::rmse)
hist(errs2)
}
