% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_loo.R
\name{crossv_loo}
\alias{crossv_loo}
\title{Leave-on-out cross-validation}
\usage{
crossv_loo(data, id = ".id")
}
\arguments{
\item{data}{A data frame}

\item{id}{Name of variable that gives each model a unique integer id.}
}
\value{
A data frame with n rows and columns \code{test} and \code{train}.
  \code{test} and \code{train} are list-columns containing
  \code{\link{resample}} objects.
}
\description{
Leave one out cross-validation. This is a convenience function
for \code{\link[modelr]{crossv_kfold}} when \code{k = nrow(data)}.
}
