% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_lpo.R
\name{crossv_lpo}
\alias{crossv_lpo}
\alias{crossv_lpo_n}
\alias{crossv_loo_n}
\alias{crossv_loo}
\title{Generate cross-validated leave-p-out test/training pairs}
\usage{
crossv_lpo(expr, p = 1L, n = NULL, extractor = NULL)

crossv_lpo_n(n, p = 1L)

crossv_loo_n(n, ...)

crossv_loo(expr, ...)
}
\arguments{
\item{expr}{An expression for the data to be sampled. This does not need to be a data frame, as \code{\link[=LazySample]{lazy_sample}} objects support or can be extended to support many classes. The unevaluated expression and its environment are captured as a \code{\link[rlang]{quosure}}.}

\item{p}{The number of elements to include in the test set. For example, specify \code{p = 1} for leave-one-out cross validation.}

\item{n}{A non-negative scalar integer representing the number of elements to sample from. If \code{n} is \code{NULL}, then \code{expr} is evaluated and called be \code{\link{idx_len}} to calculate this number. However, if \code{n} is non-\code{NULL}, then \code{expr} is not evaluated at all until samples are requested.}

\item{extractor}{A function or object coercible to a function by \code{\link[rlang]{as_function}}. See \link{LazySample} for more information.}

\item{...}{Arguments passed to \code{\link{crossv_lpo_n}}.}
}
\value{
A data frame (\code{\link[tibble]{tibble}}) with \code{combn(n, p)} rows and the following columns:
\describe{
\item{train}{A list column with objects representing the training sets. For \code{crossv_lpo} and \code{crossv_loo}, the elements are vectors of the indexes; for \code{crossv_lpo_n} and \code{crossv_loo_n}, the elements are \code{\link[=LazySample]{lazy_sample}} objects.}
\item{test}{A list column with objects representing the training sets. For \code{crossv_lpo} and \code{crossv_loo}, the elements are vectors of the indexes; for \code{crossv_lpo_n} and \code{crossv_loo_n}, the elements are \code{\link[=LazySample]{lazy_sample}} objects.}
}
}
\description{
Generate cross-validated leave-one-out or leave-p-out
test/training pairs.
}
\details{
The function \code{crossv_loo} generates leave-one-out test/training pairs.
The function \code{cross_lpo} generates the more general leave-p-out cross-validation test/training pairs.
}
\references{
\itemize{
\item{Breiman, L., Friedman, J.H., Olshen, R.A. and Stone, C.J. (1984) Classification and Regression Trees. Wadsworth.}
\item{Burman, P. (1989) A comparative study of ordinary cross-validation, v-fold cross-validation and repeated learning-testing methods. Biometrika, 76, 503–514}
\item{Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.}
\item{Efron, B. (1986) How biased is the apparent error rate of a prediction rule? Journal of the American Statistical Association, 81, 461–470.}
\item{Stone, M. (1974) Cross-validation choice and assessment of statistical predictions (with Discussion). Journal of the Royal Statistical Society, B, 36, 111–147.}
}
}
\seealso{
Other cross-validation functions: \code{\link{crossv_idx}},
  \code{\link{crossv_mc}}
}
