% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_repeated.R
\name{crossv_repeated}
\alias{crossv_repeated}
\title{Repeated k-fold cross-validation}
\usage{
crossv_repeated(data, k = 5, n = 1, id = ".id")
}
\arguments{
\item{data}{A data frame}

\item{id}{Name of variable that gives each model a unique integer id.}
}
\value{
A data frame with n rows and columns \code{test} and \code{train}.
  \code{test} and \code{train} are list-columns containing
  \code{\link{resample}} objects.
}
\description{
Repeated k-fold cross-validation.
}
