% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscv.R
\name{crossv_ts}
\alias{crossv_ts}
\alias{crossv_ts.data.frame}
\alias{crossv_ts.grouped_df}
\title{Generate Time Series Cross Validation}
\usage{
crossv_ts(data, ...)

\method{crossv_ts}{data.frame}(data, horizon = 1L, test_size = 1L,
  test_partial = FALSE, train_partial = TRUE, train_max = n,
  test_start = NULL, from = 1L, to = n, by = 1L, ...)

\method{crossv_ts}{grouped_df}(data, horizon = 1L, test_size = 1L,
  test_partial = FALSE, train_partial = TRUE, train_max = n,
  test_start = NULL, from = 1L, to = n, by = 1L, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{horizon}{Difference between the first test set observation and the last training set observation}

\item{test_size}{Size of the test set}

\item{test_partial}{If \code{TRUE}, then allows for partial test sets. If \code{FALSE},
does not allow for partial test sets.  If a number, then it is the minimum allowable
size of a test set.}

\item{train_partial}{Same as \code{test_partial}, but for the training set.}

\item{train_max}{The maximum size of the training set. This allows for a rolling window
training set, instead of using all obsservations from the start of the time series.}

\item{test_start, from, to, by}{An integer vector of the starting index values of the test set.
\code{NULL}, then these are generated from \code{seq(from, to, by)}.}
}
\description{
Also called "evaluation on a rolling forecasting origin"
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Data frame method. This rows are assumed to be ordered.

\item \code{grouped_df}: Grouped data frame method. The groups are assumed to be ordered, and
the cross validation works on groups rather than rows.
}}

\references{
\itemize{
\item{Rob J. Hyndman. \href{Cross-validation for time series}{http://robjhyndman.com/hyndsight/tscv/}. December 5, 2016.}

\item{Rob J. Hyndman. \href{Time series cross-validation: an R example}{http://robjhyndman.com/hyndsight/tscvexample/}. August 26, 2011.}
}
}
