% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_tskfold.R
\name{crossv_tskfold}
\alias{crossv_tskfold}
\title{Generate cross-validated time-series k-fold test-training pairs}
\usage{
crossv_tskfold(data, k, ...)
}
\arguments{
\item{...}{Passed to methods}
}
\description{
Generate cross-validated time-series k-fold test-training pairs
}
