% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sample.R
\name{extract_sample}
\alias{extract_sample}
\alias{extract_sample.default}
\alias{extract_sample.matrix}
\alias{extract_sample.data.frame}
\alias{extract_sample.quosure}
\title{Extract samples from objects}
\usage{
extract_sample(x, idx, ...)

\method{extract_sample}{default}(x, idx, margin = 1L, ...)

\method{extract_sample}{matrix}(x, idx, rows = TRUE, drop = FALSE, ...)

\method{extract_sample}{data.frame}(x, idx, rows = TRUE, drop = FALSE, ...)

\method{extract_sample}{quosure}(x, ...)
}
\arguments{
\item{x}{An object}

\item{idx}{An integer or character vector representing index values}

\item{...}{Arguments used by methods.}

\item{margin}{An integerish scalar. The margin (dimension) to which \code{idx} refer. For example, \code{extract(x, idx, 1)} is like \code{x[idx, ]}, \code{extract(x, idx, 2)} is like \code{x[ , idx, ...]}, and so on.}

\item{rows}{A logical scalar. If \code{TRUE}, then extract rows, else extract columns.}

\item{drop}{A logical scalar. If \code{TRUE}, then coerce to the lowest possible dimension. See \code{\link{drop}}.}
}
\description{
The generic function \code{\link{extract_sample}} is the function that
extracts the elements specified by \code{idx} from the the data in a
\code{lazy_sample} object.
}
\details{
It is largely for internal use in order to abstract the details for
extracting elements from data frames, vectors, matrices, arrays, etc.
However, it is exported so that the user can add methods for new objects
or change the default behavior. The \code{extractor} field in \code{lazy_sample}
objects and the \code{extractor} argument in many functions will override
this function.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: The default method is equivalent to \code{x[idx]} when there are no dimensions in \code{x}. When there are dimensions in \code{x}, then \code{idx} refers to the \code{margin} dimension.

\item \code{matrix}: The default method for matrices extracts rows by default and does not drop dimensions.

\item \code{data.frame}: The data frame method extracts rows by default, and does not drop dimensions.

\item \code{quosure}: For quosures, the method will evaluate the expression and then call the next method on the result.
}}

