% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout_idx.R
\name{holdout_idx}
\alias{holdout_idx}
\alias{holdout_idx.grouped_df}
\title{Generate cross-validation test-training sets manuall}
\usage{
holdout_idx(data, train, test, ...)

\method{holdout_idx}{grouped_df}(data, train, test, ...)
}
\arguments{
\item{data}{A data frame}

\item{train}{A list of integer vectors or an integer vector, with the indices of the training set.
If \code{train} is \code{NULL}, then the training set is the complement
the test set.}

\item{test}{A list of integer vectors or an integer vector, with the indices of the test set.
If \code{test} is \code{NULL}, then the test set is the complement of
the training set.}

\item{...}{Arguments passed to methods}
}
\value{
A data frame with
}
\description{
Generate cross-validation test-training sets from manually specified
indices of observations in the test and and training sets.
}
\section{Methods (by class)}{
\itemize{
\item \code{grouped_df}: Partition a grouped data frame into test and training
sets by group.
}}

\examples{
library("dplyr")

# test-training pairs can be specified by test sets
holdout_idx(mtcars, train = list(1:5, 6:10))
# or by training sets
holdout_idx(mtcars, train = list(1:20, 10:32))
# or by both (and they don't need to partition the observations)
holdout_idx(mtcars,
          train = list(1:20, 10:32),
          test = list(21:32, 1:9))

# with grouped data frames, the indices refer to groups
holdout_idx(group_by(mtcars, mpg), train = list(1L))$train
}
