% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{is.resample}
\alias{is.resample}
\title{Is it a resample object?}
\usage{
is.resample(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks whether an object is an \code{\link{resample}}
object.
}
