% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{is_resample}
\alias{is_resample}
\title{Is it a resample object?}
\usage{
is_resample(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks whether an object inherits from \code{\link[modelr]{resample}}.
}
