% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jack_bias}
\alias{jack_bias}
\title{Jackknife bias estimate}
\usage{
jack_bias(jack)
}
\arguments{
\item{jack}{Object from \code{\link{jack_stat}}.}
}
\value{
Jackknife estimate of bias.
}
\description{
Jackknife bias estimate
}
