% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jack_se}
\alias{jack_se}
\title{Jackknife standard error}
\usage{
jack_se(jack)
}
\arguments{
\item{jack}{Object from \code{\link{jack_stat}}.}
}
\value{
Jackknife estimate of standard error.
}
\description{
Jackknife standard error
}
