% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\alias{jackknife_n}
\title{Generate Jackknife Subsamples}
\usage{
jackknife(expr, p = 1L, n = NULL, extractor = NULL)

jackknife_n(n, p = 1L)
}
\arguments{
\item{expr}{An expression for the data to be sampled. This does not need to be a data frame, as \code{\link[=LazySample]{lazy_sample}} objects support or can be extended to support many classes. The unevaluated expression and its environment are captured as a \code{\link[rlang]{quosure}}.}

\item{p}{The number of elements to delete. \code{p = 1} is the ordinary
(delete-one) jackknife, while \code{p > 1} is the block (delete-\eqn{p})
jackknife.}

\item{n}{A non-negative scalar integer representing the number of elements to sample from. If \code{n} is \code{NULL}, then \code{expr} is evaluated and called be \code{\link{idx_len}} to calculate this number. However, if \code{n} is non-\code{NULL}, then \code{expr} is not evaluated at all until samples are requested.}

\item{extractor}{A function or object coercible to a function by \code{\link[rlang]{as_function}}. See \link{LazySample} for more information.}

\item{...}{Arguments passed to methods}
}
\value{
data frame (\code{\link[tibble]{tibble}}) with \code{combn(n, p)} (\code{n} if \code{p = 1}) rows and the following columns.
\describe{
\item{sample}{A list column with objects representing the training sets. For jackknife, the elements are vectors of the indexes; for jackknife_n, the elements are \code{\link[=LazySample]{lazy_sample}} objects.}
}
}
\description{
Generate jackknife subsamples. For data with \eqn{n} elements,
the jackknife generates \eqn{n} subsamples, each with one element
deleted. The generalized (delete-\eqn{p}, block) jackknife generates
subsamples by deleting \code{p} observations.
}
\references{
\itemize{
\item{Davison, A. C. & Hinkley, D. V. (1997) Bootstrap Methods and Their Applications. Cambridge University Press, Cambridge. ISBN 0-521-57391-2}
\item{Tukey, J. W. (1958). "Bias and confidence in not quite large samples". The Annals of Mathematical Statistics.  \href{https://dx.doi.org/10.1214/aoms/1177706647}{doi:10.1214/aoms/1177706647}.}
\item{Efron, B.; Stein, C. (May 1981). "The Jackknife Estimate of Variance". The Annals of Statistics. \href{https://dx.doi.org/10.1214/aos/1176345462}{doi:10.1214/aos/1176345462}.}
}
}
