% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_sample.R
\name{lazy_sample}
\alias{lazy_sample}
\alias{lazy_sample_lst}
\alias{lazy_sample_idx}
\title{Create lazy sample objects}
\usage{
lazy_sample(expr, idx, extractor = NULL)

lazy_sample_lst(expr, idx, extractor = NULL)

lazy_sample_idx(expr, idx, extractor = NULL)
}
\arguments{
\item{expr}{A quosure}

\item{idx}{A list of integer or character vector with the indexes in
each sample.}

\item{extractor}{See \code{\link{lazy_sample}}.}
}
\value{
\describe{
\item{lazy_sample}{A \code{LazySample} object.}
\item{lazy_sample_lst}{A list of \code{LazySample} objects with length equal to the length of \code{idx}.}
\item{lazy_sample_idx}{A data frame with one column and rows equal to the length of \code{idx}.
The single column is named \code{sample} and is a list of \code{LazySample} objects.}
}
}
\description{
The function \code{lazy_sample} is a convience function to create
\code{LazySample} objects. The function \code{lazy_sample_lst} creates a list
of lazy sample objects with the same expression, but multiple indexes.
The function \code{lazy_sample_idx} creates a data frame with a list-column of \code{LazySample} objects.
}
