% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{loo_cv}
\alias{loo_cv}
\title{Leave-one-out cross-validation}
\usage{
loo_cv(data, y, model_fit, predict_fun, loss, ...)
}
\arguments{
\item{data}{Data frame or list of inputs.}

\item{y}{Response vector (numeric or factor).}

\item{model_fit}{Function \code{model_fit(data_train, y_train, ...)}
returning a fitted model object.}

\item{predict_fun}{Function \code{predict_fun(fit, data_test)} returning
predicted values.}

\item{loss}{Function \code{loss(y_true, y_pred)} returning mean loss.}

\item{...}{Additional arguments passed to \code{model_fit}.}
}
\description{
Leave-one-out cross-validation
}
