% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{perm_test}
\alias{perm_test}
\title{Two-sample permutation test}
\usage{
perm_test(
  x,
  y,
  stat = function(x, y) mean(x) - mean(y),
  B = 9999L,
  alternative = c("two.sided", "less", "greater"),
  exact = FALSE
)
}
\arguments{
\item{x}{Numeric vector for group 1.}

\item{y}{Numeric vector for group 2.}

\item{stat}{Function \code{stat(x, y)} returning numeric scalar.
Default is difference in means, \code{mean(x) - mean(y)}.}

\item{B}{Number of random permutations.}

\item{alternative}{One of \code{"two.sided"}, \code{"less"}, \code{"greater"}.}

\item{exact}{Logical; if \code{TRUE} and \code{length(x)+length(y)} is small,
enumerate all permutations to obtain an exact p-value.}
}
\value{
A list with components
\describe{
\item{stat_obs}{Observed test statistic.}
\item{perm_stats}{Vector of permutation statistics.}
\item{p_value}{Estimated p-value (ASL).}
}
}
\description{
Perform a permutation test for difference between two samples using
an arbitrary test statistic.
}
