% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\alias{permute.data.frame}
\alias{permute.grouped_df}
\title{Generate permutation samples}
\usage{
permute(x, ...)

\method{permute}{data.frame}(x, k = 1L, ...)

\method{permute}{grouped_df}(x, k = 1L, stratify = FALSE, ...)
}
\arguments{
\item{x}{A data frame}

\item{...}{Passed to methods}

\item{k}{The number of permutations to generate. If \code{k = Inf}, then
the full set of permutations is generated.}

\item{stratify}{If \code{TRUE}, the permute observations within each group.
Otherwise permute the groups themselves.}
}
\value{
A data frame with \code{k} rows and the following columns:
\describe{
\item{sample}{A list of \code{\link{resample}} objects.}
\item{.id}{Character vector. ID of the sample.}
}
}
\description{
Generate permutation samples
}
