% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_pois.R
\name{pois_bootstrap}
\alias{pois_bootstrap}
\alias{resample_pois_bootstrap}
\title{Generate Poisson Bootstrap Replicates}
\usage{
pois_bootstrap(data, n, size = NULL, wt = NULL)

resample_pois_bootstrap(data, size = NULL, wt = NULL)
}
\arguments{
\item{data}{a data frame}

\item{n}{Number of replicates to draw}

\item{size}{Expected size of each replicate sample}

\item{wt}{Observation weights. If \code{NULL}, all observations
are weighted equally.}
}
\description{
Generates replicates from a Poisson bootstrap, which
samples each observation independently from a Poisson distribtuion
with \expr{lambda = 1} if all obs have equal probability.
}
