% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Create a lazy resample object}
\usage{
resample(data, idx)
}
\arguments{
\item{data}{A data frame}

\item{idx}{An integer vector}
}
\value{
A \code{resample} object, a list with two elements: \code{data}
  and \code{idx}.
}
\description{
This is a slightly modified version of \code{\link[modelr]{resample}}.
It has more tests for the validity of the indexes, but also coerces
numeric \code{idx} to integers.
}
