% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_df.R
\name{resample_data}
\alias{resample_data}
\title{Generate subsamples from indices}
\usage{
resample_data(data, ...)
}
\arguments{
\item{data}{An object}

\item{...}{Arguments passed to methods}

\item{samples}{A list of integer vectors or an integer vector, with the indices in each subsample.}
}
\value{
A data frame with rows for each subsample and the following columns:
\describe{
\item{train}{A list of \code{\link[modelr]{resample}} objects. Training sets.}
\item{test}{A list of \code{\link[modelr]{resample}} objects. Test sets.}
\item{.id}{An integer vector of identifiers}
}
}
\description{
Generate subsamples from manually specified indices of observations in each
subsample.
}
\examples{
library("dplyr")
# specify subsamples by a list of indices
resample_df(mtcars, samples = c(1:5, 5:10, 15:20))
resample_df(mtcars, samples = c(1:10, 1:15, 5:30))
# with grouped data frames, indices refer to groups
resample_df(group_by(mtcars, mpg), samples = list(c(1:2)))$sample
}
