% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_idx.R
\name{resample_idx}
\alias{resample_idx}
\alias{resample_idx.default}
\alias{resample_idx.grouped_df}
\alias{resample_idx.grouped_df}
\title{Generate subsamples from indices}
\usage{
resample_idx(data, ...)

\method{resample_idx}{default}(data, samples, ...)

\method{resample_idx}{grouped_df}(data, samples, ...)

\method{resample_idx}{grouped_df}(data, samples, ...)
}
\arguments{
\item{data}{An object}

\item{...}{Arguments passed to methods}

\item{samples}{A list of integer vectors or an integer vector, with the indices in each subsample.}
}
\value{
A data frame with rows for each subsample and the following columns:
\describe{
\item{sample}{A list of \code{\link[modelr]{resample}} objects. Training sets.}
\item{.id}{An integer vector of identifiers.}
}
}
\description{
Generate subsamples from manually specified indices of observations in each
subsample.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Generate subsamples of rows of data frames.

\item \code{grouped_df}: Generate subsamples of groups of grouped data frames.

\item \code{grouped_df}: Generate subsamples of groups of grouped data frames.
}}

\examples{
resample(mtcars, 1:10)
}
