% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{resample_jackknife}
\alias{resample_jackknife}
\title{Generate a jacknkife replicate}
\usage{
resample_jackknife(data, i)
}
\arguments{
\item{data}{A data frame}

\item{i}{The row index number to exclude}
}
\value{
A \code{\link[modelr]{resample}} object.
}
\description{
Generate a jacknife replicate, which includes all rows of \code{data},
but a given row \code{i}.
}
