% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resample_list}
\alias{resample_list}
\title{Create a list of resample objects}
\usage{
resample_list(data, idxs)
}
\arguments{
\item{data}{A data frame}

\item{idx}{A list of integer vectors of indexes}
}
\value{
A list of \code{\link[modelr]{resample}} objects.
}
\description{
Create a list of resample objects from a data frame and a list
of indexes.
}
