% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{resample_permutation}
\alias{resample_permutation}
\title{Generate a permutation replicate}
\usage{
resample_permutation(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A \code{\link[modelr]{resample}} object.
}
\description{
Generate a permutation replicate, which randomly shuffles the rows of
the data frame.
}
