% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_split.R
\name{resample_split}
\alias{resample_split}
\alias{resample_groups}
\title{Generate resample objects from splits}
\usage{
resample_split(data, test = NULL, train = NULL)

resample_groups(data, test = NULL, train = NULL)
}
\arguments{
\item{data}{A data table}

\item{test, train}{Integer vectors with the indexes of the test
and training samples. One of test or train must be specified.}
}
\value{
A named list of two \code{\link[modelr]{resample}} objects
  for the "test" and "train" sets.
}
\description{
Generate resample objects from splits
}
