% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_ts.R
\name{resample_ts_bootstrap}
\alias{resample_ts_bootstrap}
\title{Generate a time-series bootstrap replicate}
\usage{
resample_ts_bootstrap(data, size = NULL, len = NULL, sim = c("fixed",
  "geom"), endcorr = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{size}{Length of the simulated time series in the replicate.}

\item{len}{If \code{sim} is \code{"fixed"} then \code{len} is the fixed block
length used in generating the replicate time series.
If \code{sim} is \code{"geom"} then \code{len} is the mean of the
geometric distribution used to generate the block lengths.
\code{length} should be a positive integer less than the number of rows
of \code{data}.}

\item{sim}{The type of simulation required to generate the replicate time series.
The possible input values are "model" (model based resampling), "fixed" (block resampling with fixed block lengths of l), "geom" (block resampling with block lengths having a geometric distribution with mean l)}

\item{A}{logical variable indicating whether end corrections are to be applied when sim is "fixed". When sim is "geom", endcorr is automatically set to TRUE; endcorr is not used when sim is "model" or "scramble".}
}
\description{
Generate a time-series bootstrap replicate
}
\seealso{
\code{\link[boot]{tsboot}}
}
