% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{resample_window}
\alias{resample_window}
\title{Generate a window}
\usage{
resample_window(data, idx, width, partial = TRUE, align = c("left", "right",
  "center"))
}
\arguments{
\item{data}{A data frame}

\item{idx}{An integer index}

\item{width}{The width of the window}

\item{partial}{If \code{TRUE}, and the window is not fully inside the data,
then a truncated window is returned. Otherwise, \code{NULL} is returned.}

\item{align}{Alignment of the window relative to \code{idx}}
}
\value{
A \code{resample} object or \code{NULL}, if the window falls outside
  the data.
}
\description{
Generate a \code{\link[modelr]{resample}} object for a window.
}
