% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resampler-internal}
\alias{resampler-internal}
\alias{resample_idx_len}
\alias{resample_idx_len.default}
\alias{resample_idx_len.matrix}
\alias{resample_idx_len.array}
\alias{resample_idx_len.data.frame}
\alias{resample_idx_len.resample}
\alias{resample_extract}
\alias{resample_extract.default}
\alias{resample_extract.matrix}
\alias{resample_extract.array}
\alias{resample_idx_len.grouped_df}
\title{Interal methods for \code{resampler} objects}
\usage{
resample_idx_len(x, ...)

\method{resample_idx_len}{default}(x, ...)

\method{resample_idx_len}{matrix}(x, ...)

\method{resample_idx_len}{array}(x, margin = 1L, ...)

\method{resample_idx_len}{data.frame}(x, ...)

\method{resample_idx_len}{resample}(x, ...)

resample_extract(x, ...)

\method{resample_extract}{default}(x, i, ...)

\method{resample_extract}{matrix}(x, i, margin = 1L, ...)

\method{resample_extract}{array}(x, margin = 1L, ...)

\method{resample_idx_len}{grouped_df}(x, groups = TRUE, ...)
}
\arguments{
\item{x}{An object (the \code{data} element of a \code{resampler} object)}

\item{...}{Additional arguments (currently unused)}

\item{margin}{Integer. Margin of matrix or arrary to use for indices.}

\item{i}{Vector. Indices.}
}
\value{
\describe{
\item{resamplr_idx_len}{An integer of length 1.}
\item{resamplr_extract}{An object of the same class as \code{x}}
}
}
\description{
These are methods used to handle different types of \code{data} elements
in \code{resample} objects.
Instead of subclassing \code{resample} objects, it is easier to define
methods for the several key tasks needed.
}
