% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Generate rolling windows}
\usage{
roll(data, width, by = 1L, partial = FALSE, align = c("center", "left",
  "right"))
}
\arguments{
\item{data}{A data frame}

\item{width}{The width of the window}

\item{by}{Start windows at every by-th point rather than every point.}

\item{partial}{If \code{TRUE}, and the window is not fully inside the data,
then a truncated window is returned. Otherwise, \code{NULL} is returned.}

\item{align}{Alignment of the window relative to \code{idx}}
}
\value{
A data frame with four columns
\describe{
\item{window}{A list column of \code{\link[modelr]{resample}} objects}
\item{.id}{integer: the index position of the window}
\item{.width}{integer: the width of the window}
\item{.align}{character: alignment of the window. "l" for left,
"c" for center, "r" for right}
}
}
\description{
Generate \code{\link[modelr]{resample}} objects for rolling windows.
}
\seealso{
The \code{\link[zoo]{rollapply}} function in \pkg{zoo}
}
