% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbootstrap.R
\name{tsbootstrap}
\alias{tsbootstrap}
\alias{tsbootstrap.data.frame}
\alias{tsbootstrap.grouped_df}
\title{Generate time-series bootstrap replicates}
\usage{
tsbootstrap(data, ...)

\method{tsbootstrap}{data.frame}(data, R = 1L, size = 1L, m = nrow(data),
  type = "fixed", ...)

\method{tsbootstrap}{grouped_df}(data, R = 1L, size = 1L, type = "fixed",
  m = dplyr::n_groups(data), ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{R}{Number of bootstrap replicates}

\item{size}{If sim is \code{"fixed"} then \code{size} is the fixed block
length used in generating the replicate time series. If sim is
\code{"geom"} then \code{size} is the mean of the geometric distribution
used to generate the block lengths. \code{size} should be a positive
integer less than the number of rows (\code{data.frame}) or groups
\code{grouped_df} in \code{data}.}

\item{m}{Lengths of time-series replicates}

\item{type}{Type of simulation used to generate the replicate time series.
The possible input values are "fixed" (block resampling with fixed block
lengths of \code{size}), "geom" (block resampling with block lengths
having a geometric distribution with mean \code{size}).}
}
\value{
A data frame with \code{R} rows and the following columns:
\describe{
\item{train}{A list of \code{\link[modelr]{resample}} objects. Training sets.}
\item{test}{A list of \code{\link[modelr]{resample}} objects. Test sets.}
\item{.id}{An integer vector of identifiers}
}
}
\description{
Generate \code{R} bootstrap replicates for time series.
The replicate time series can be generated using either fixed or random
block lengths.
}
\details{
If \code{type} is \code{"fixed"} then each replicate time series is found
by taking blocks of length \code{size}, from the original time series and
putting them end-to-end until a new series of length \code{m} is created.
When \code{type} is \code{"geom"}, a similar approach is taken except
that the block lengths are sampled from a geometric distribution with mean
\code{size}.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Generate time series replicates using blocks of
rows in the data frame.

\item \code{grouped_df}: Generate time series replicates using blocks of
groups in the grouped data frame.
}}

\references{
\itemize{
\item{Kunsch, H.R. (1989) The jackknife and the bootstrap for general stationary observations. \emph{Annals of Statistics}}
\item{Davison, A.C. and Hinkley, D.V. (1997) \emph{Bootstrap Methods and Their Application}. Cambridge University Press.}
\item{Politis, D.N. and Romano, J.P. (1994) "The stationary bootstrap. Journal of the American Statistical Association."}
\item{Angelo Canty and Brian Ripley (2016). boot: Bootstrap R (S-Plus) Functions. R package version 1.3-18.}
}
}
\seealso{
The \pkg{boot} function \code{\link[boot]{tsboot}}, from which this
  function is derived.
}
