% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_balanced.R
\name{balanced_bootstrap}
\alias{balanced_bootstrap}
\title{Generate n balanced bootstrap replicates}
\usage{
balanced_bootstrap(data, n, strata = FALSE, clusters = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{n}{number of replicates}

\item{strata}{strata}

\item{clusters}{clusters}
}
\description{
Generate n balanced bootstrap replicates
}
