% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_bootstrap.R
\name{balanced_bootstrap}
\alias{balanced_bootstrap}
\alias{balanced_bootstrap.data.frame}
\alias{balanced_bootstrap.grouped_df}
\title{Generate balanced bootstrap replicates}
\usage{
balanced_bootstrap(data, ...)

\method{balanced_bootstrap}{data.frame}(data, R = 1L, ...)

\method{balanced_bootstrap}{grouped_df}(data, R = 1L, stratify = FALSE, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{R}{Number of replicates}

\item{stratify}{If \code{TRUE}, resample within groups, otherwise,
resample by groups instead of rows.}
}
\value{
A data frame with \code{R} rows and rows the following columns:
\describe{
\item{train}{A list of \code{\link{resample}} objects. Training sets.}
\item{test}{A list of \code{\link{resample}} objects. Test sets.}
\item{.id}{An integer vector of identifiers}
}
}
\description{
A balanced bootstrap generates \code{R} replicates in which each element
appears \code{R} times.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Balanced bootstrap of rows in a data frame.

\item \code{grouped_df}: Balanced bootstrap either by group, or
within groups (if \code{stratify = TRUE}).
}}

\examples{
# example from modelr::bootstrap
library("purrr")
library("dplyr")
boot <- balanced_bootstrap(mtcars, 100)

models <- map(boot$sample, ~ lm(mpg ~ wt, data = .))
tidied <- map_df(models, broom::tidy, .id = "id")

hist(subset(tidied, term == "wt")$estimate)
hist(subset(tidied, term == "(Intercept)")$estimate)
}
\references{
\itemize{
\item{Gleason, John 1988. "Algorithms for Balanced Bootstrap Simulations." \emph{The American Statistician}
\href{https://dx.doi.org/10.2307/2685134}{doi:10.2307/2685134}.}
\item{Davison, A.C., Hinkley, D. V., and Schechtman, E. 1986.
"Efficient Bootstrap Simulation." \emph{Biometrika}.}
\item{Angelo Canty and Brian Ripley (2016). boot: Bootstrap R (S-Plus) Functions. R package version 1.3-18.}
}
}
