% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap_n}
\title{Generate bootstrap replicates}
\usage{
bootstrap(expr, R = 1L, weights = NULL, bayesian = FALSE, m = NULL,
  replace = TRUE, n = NULL, extractor = NULL)

bootstrap_n(n, R = 1L, weights = NULL, bayesian = FALSE, m = NULL,
  replace = TRUE)
}
\arguments{
\item{expr}{A quosure}

\item{R}{Number of replicates to sample}

\item{weights}{Observation level weights to use for a weighted bootstrap. If \code{bayesian = TRUE}, then the sum of these weights will adjust the concentration of the Dirichlet prior.}

\item{bayesian}{If \code{TRUE}, a Bayesian bootstrap is used.}

\item{m}{Number of observations in the replicates. This can be set to values \code{m != n} for sub-sampling or oversampling.}

\item{replace}{Logical scalar. If \code{TRUE}, then sample with replacement, otherwise sample without replacement.}

\item{n}{Number of elements to sample from. If \code{NULL}, then
\code{n = len_idx(expr)}.}

\item{extractor}{See \code{\link{lazy_sample}}.}
}
\description{
Generate bootstrap replicates. In addition to the ordinary bootstrap,
it supports weighted bootstrap (\code{weights}),
Bayesian bootstrap (\code{bayes}), cluster bootstrap (if \code{data} is a
grouped data frame and \code{groups = TRUE}), and stratified bootstrap
(if \code{data} is grouped data frame and \code{stratify = TRUE}).
}
\references{
\itemize{
\item{Angelo Canty and Brian Ripley (2016). boot: Bootstrap R (S-Plus) Functions. R package version 1.3-18.}
\item{Davison, A. C. & Hinkley, D. V. (1997) Bootstrap Methods and Their Applications. Cambridge University Press, Cambridge. ISBN 0-521-57391-2}
}
}
\seealso{
The \pkg{boot} function \code{\link[boot]{boot}} which is the
  canonical R bootstrap implementation.
  The \pkg{modelr} function \code{\link[modelr]{bootstrap}} also implements a lazy bootstrap.
}
