% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap.data.frame}
\alias{bootstrap.grouped_df}
\title{Generate bootstrap replicates}
\usage{
bootstrap(data, ...)

\method{bootstrap}{data.frame}(data, k = 1L, weights = NULL,
  bayes = FALSE, ...)

\method{bootstrap}{grouped_df}(data, k = 1L, stratify = FALSE,
  groups = TRUE, weights = NULL, weight_groups = TRUE,
  weight_within = TRUE, bayes = FALSE, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Passed to methods}

\item{k}{Number of replicates to sample}

\item{weights}{Observation level weights. This must be the name of a column
in \code{data}.}

\item{bayes}{If \code{TRUE}, a Bayesian bootstrap is used.}

\item{stratify}{Resample within groups (stratified bootstrap)}

\item{groups}{Resample groups (clustered bootstrap)}

\item{weight_groups}{If \code{TRUE} and \code{groups = TRUE},
then sample groups using group-level weights.
If \code{weights} is \code{NULL}, then weight groups by the number
of observations in each. If \code{weights}, is not-\code{NULL}, then
use the sum of the observation level weights in each group as the
group-level weight.}

\item{weight_within}{If \code{TRUE} and \code{stratify = TRUE}, then
use observation level weights to sample observations within each group.}
}
\value{
A data frame with \code{k} rows and columns:
\describe{
\item{\code{sample}}{A list of \code{\link{resample}} objects with
with the bootstrap replicates.}
\item{\code{.id}}{An integer vector with replicate ids}
}
}
\description{
Generate bootstrap replicates. In addition to the ordinary bootstrap,
it supports weighted bootstrap (\code{weights}),
Bayesian bootstrap (\code{bayes}), cluster bootstrap (if \code{data} is a
grouped data frame and \code{groups = TRUE}), and stratified bootstrap
(if \code{data} is grouped data frame and \code{stratify = TRUE}).
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Bootstrap rows in a data frame.

\item \code{grouped_df}: Bootstraps a grouped data data frame. Allows for
bootstrapping groups (clustered bootstrap) if \code{groups = TRUE}, and
bootstrapping within groups (stratified bootstrap) if
\code{stratify = TRUE}.
}}

\examples{
# example from modelr::bootstrap
library("purrr")
library("dplyr")
boot <- balanced_bootstrap(mtcars, 100)

models <- map(boot$sample, ~ lm(mpg ~ wt, data = .))
tidied <- map_df(models, broom::tidy, .id = "id")

hist(subset(tidied, term == "wt")$estimate)
hist(subset(tidied, term == "(Intercept)")$estimate)
}
