% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Generate n bootstrap replicates}
\usage{
bootstrap(data, n, ..., id = ".id")
}
\arguments{
\item{data}{A data frame}

\item{n}{Number of bootstrap replicates to generate}

\item{...}{Passed to \code{\link{resample_bootstrap}}}

\item{id}{Name of variable that gives each model a unique integer id}
}
\value{
A data frame with two columns and \code{n} rows
\describe{
\item{strap}{A list column of \code{\link[modelr]{objects}}}
\item{id}{The replicate identifier}
}
}
\description{
Generate n bootstrap replicates
}
