% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_df.R
\name{crossv_df}
\alias{crossv_df}
\alias{crossv_df.data.frame}
\alias{crossv_df.grouped_df}
\title{Generate cross-validation test-training sets manually}
\usage{
crossv_df(data, ...)

\method{crossv_df}{data.frame}(data, train = NULL, test = NULL, ...)

\method{crossv_df}{grouped_df}(data, train = NULL, test = NULL, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{train}{A list of integer vectors, with the indices of the training set.
If \code{train} is \code{NULL}, then the training set is the complement of
the test set.}

\item{test}{A list of integer vectors, with the indices of the test set.
If \code{test} is \code{NULL}, then the test set is the complement of
the training set.}
}
\description{
Generate cross-validation test-training sets from manually specified
indices of observations in the test and and training sets.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Partition a data frame into test and training sets by row.

\item \code{grouped_df}: Partition a grouped data frame into test and training
sets by group.
}}

