% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv.R
\name{crossv_idx}
\alias{crossv_idx}
\title{Generate cross-validation test-training sets manually}
\usage{
crossv_idx(data, train, test, ...)
}
\arguments{
\item{data}{A data frame}

\item{train}{A list of integer vectors or an integer vector, with the indices of the training set.
If \code{train} is \code{NULL}, then the training set is the complement
the test set.}

\item{test}{A list of integer vectors or an integer vector, with the indices of the test set.
If \code{test} is \code{NULL}, then the test set is the complement of
the training set.}

\item{...}{Arguments passed to methods}
}
\description{
Generate cross-validation test-training sets from manually specified indices of observations in the test and and training sets.
}
