% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_idx.R
\name{crossv_idx}
\alias{crossv_idx}
\title{Generate cross-validation test-training sets}
\usage{
crossv_idx(expr, train, test, extractor = NULL)
}
\arguments{
\item{expr}{A quosure}

\item{train}{A list of integer or character vectors}

\item{test}{A list of integer or character vector}

\item{extractor}{A function or object coercible to a function by \code{\link[rlang]{as_function}}. See \link{LazySample} for more information.}
}
\value{
A data frame with <%= numrows %> rows and the following columns:
\describe{
\item{train}{A list column with objects representing the training sets. The list elements are \code{\link{lazy_sample}} objects.}
\item{test}{A list column with objects representing the training sets. The list elements are \code{\link{lazy_sample}} objects.}
}
}
\description{
Generate cross-validation test/training pairs by specifying the indices
in the elements in the test and training sets.
This function provides a manual method for generating data frames with
with cross-validation \code{resample} objects, while other functions,
such as \code{\link{crossv_kfold}}, algorithmically handle generate the
indexes in the test- and training sets.
}
\examples{
# specify test and training partitions
test <- list(1:10, 10:20, 30:nrow(mtcars))
train <- purrr::map(test, setdiff, y = test)
crossv_idx(mtcars, train = train, test = test)
}
\references{
\itemize{
\item{Breiman, L., Friedman, J.H., Olshen, R.A. and Stone, C.J. (1984) Classification and Regression Trees. Wadsworth.}
\item{Burman, P. (1989) A comparative study of ordinary cross-validation, v-fold cross-validation and repeated learning-testing methods. Biometrika, 76, 503–514}
\item{Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.}
\item{Efron, B. (1986) How biased is the apparent error rate of a prediction rule? Journal of the American Statistical Association, 81, 461–470.}
\item{Stone, M. (1974) Cross-validation choice and assessment of statistical predictions (with Discussion). Journal of the Royal Statistical Society, B, 36, 111–147.}
}
}
\seealso{
Other cross-validation functions: \code{\link{crossv_lpo}},
  \code{\link{crossv_mc}}
}
