% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold.R
\name{crossv_kfold}
\alias{crossv_kfold}
\alias{crossv_kfold.data.frame}
\alias{crossv_kfold.grouped_df}
\title{Generate cross-validated K-fold test-training pairs}
\usage{
crossv_kfold(data, K, ...)

\method{crossv_kfold}{data.frame}(data, K = 5L, shuffle = TRUE, ...)

\method{crossv_kfold}{grouped_df}(data, K = 5L, shuffle = TRUE,
  stratify = FALSE, ...)
}
\arguments{
\item{data}{A data frame}

\item{K}{The number of folds}

\item{...}{Arguments passed to methods}

\item{shuffle}{If \code{TRUE}, randomly assign observations to folds.
Otherwise, observations are sequentially assigned to folds.}

\item{stratify}{If \code{TRUE}, within each group observations are split
into folds, and those folds combined. If \code{FALSE}, groups are assigned
into folds.}
}
\value{
A data frame with \code{K} rows and the following columns:
\describe{
\item{sample}{A list of \code{\link[modelr]{resample}} objects. Training sets.}
\item{.id}{An integer vector of identifiers.}
}
}
\description{
Generate cross-validated K-fold test-training pairs. In addition to ordinary
K-fold cross-validation, this supports stratified K-fold cross validation
if \code{data} is a grouped data frame and \code{stratify = TRUE}, and
Group K-fold if \code{data} is a grouped data frame and \code{stratify = FALSE}.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Splits rows in a data frame into folds.

\item \code{grouped_df}: Partitions rows within each group of a grouped data frame
into folds if \code{stratify = FALSE}. This ensures that the test and training
sets will have approximately equal proportions of each group. If \code{stratify = TRUE},
then the groups are partitioned into folds.
}}

\examples{
# Example originally from modelr::crossv_mc
library("purrr")
library("dplyr")

# 5-fold cross-validation
cv1 <- crossv_kfold(mtcars, K = 5)
models <- map(cv1$train, ~ lm(mpg ~ wt, data = .))
summary(map2_dbl(models, cv1$test, modelr::rmse))

# k-fold by group
cv2 <- crossv_kfold(group_by(mtcars, cyl), K = 2)
models <- map(cv2$train, ~ lm(mpg ~ wt, data = .))
summary(map2_dbl(models, cv2$test, modelr::rmse))

# stratified k-fold
cv3 <- crossv_kfold(group_by(mtcars, am), K = 3, stratify = TRUE)
models <- map(cv3$train, ~ lm(mpg ~ wt, data = .))
summary(map2_dbl(models, cv3$test, modelr::rmse))
}
\references{
\itemize{
\item{Breiman, L., Friedman, J.H., Olshen, R.A. and Stone, C.J. (1984) Classification and Regression Trees. Wadsworth.}
\item{Burman, P. (1989) A comparative study of ordinary cross-validation, v-fold cross-validation and repeated learning-testing methods. Biometrika, 76, 503–514}
\item{Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.}
\item{Efron, B. (1986) How biased is the apparent error rate of a prediction rule? Journal of the American Statistical Association, 81, 461–470.}
\item{Stone, M. (1974) Cross-validation choice and assessment of statistical predictions (with Discussion). Journal of the Royal Statistical Society, B, 36, 111–147.}
}
}
\seealso{
This function has more features than the \pkg{modelr} function
  \code{\link[modelr]{crossv_kfold}}.
}
