% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_kfold.R
\name{crossv_kfold}
\alias{crossv_kfold}
\title{Generate cross-validated k-fold test-training pairs}
\usage{
crossv_kfold(data, k, ...)
}
\arguments{
\item{...}{Passed to methods}
}
\description{
Generate cross-validated k-fold test-training pairs
}
