% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold.R
\name{crossv_kfold}
\alias{crossv_kfold}
\alias{crossv_kfold.data.frame}
\alias{crossv_kfold.grouped_df}
\title{Generate cross-validated k-fold test-training pairs}
\usage{
crossv_kfold(data, k, ...)

\method{crossv_kfold}{data.frame}(data, k = 5, shuffle = TRUE, ...)

\method{crossv_kfold}{grouped_df}(data, k = 5, shuffle = TRUE,
  stratify = FALSE, ...)
}
\arguments{
\item{data}{A data frame}

\item{k}{The number of folds}

\item{...}{Passed to methods}

\item{shuffle}{If \code{TRUE}, randomly assign observations to folds.
Otherwise, observations are sequentially assigned to folds.}

\item{stratify}{If \code{TRUE}, within each group observations are split
into folds, and those folds combined. If \code{FALSE}, groups are assigned
into folds.}
}
\value{
A data frame with \code{k} rows and the following columns:
\describe{
\item{train}{A list of \code{\link{resample}} objects. Training sets.}
\item{test}{A list of \code{\link{resample}} objects. Test sets.}
\item{.id}{Character vector. ID of the fold.}
}
}
\description{
Generate cross-validated k-fold test-training pairs
}
