% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_kfold.R
\name{crossv_kfold}
\alias{crossv_kfold}
\alias{crossv_kfold_n}
\title{Generate cross-validated K-fold test-training pairs}
\usage{
crossv_kfold(expr, K, ...)

crossv_kfold_n(n, K = 5L, shuffle = TRUE)
}
\arguments{
\item{expr}{An expression for the data to be sampled. This does not need to be a data frame, as \code{\link[=LazySample]{lazy_sample}} objects support or can be extended to support many classes. The unevaluated expression and its environment are captured as a \code{\link[rlang]{quosure}}.}

\item{K}{An integer scalar. The number of cross-validation folds.}

\item{...}{Arguments passed to methods}

\item{n}{A non-negative scalar integer representing the number of elements to sample from. If \code{n} is \code{NULL}, then \code{expr} is evaluated and called be \code{\link{idx_len}} to calculate this number. However, if \code{n} is non-\code{NULL}, then \code{expr} is not evaluated at all until samples are requested.}

\item{shuffle}{A logical scalar. If \code{TRUE}, then the eleents are randomly shuffled prior to partitioning into folds.}
}
\value{
A data frame with \code{K} rows and the following columns:
\describe{
\item{train}{A list column with objects representing the training sets. For \code{crossv_kfold}, the elements are vectors of the indexes; for \code{crossv_kfold_n}, the elements are \code{\link[=LazySample]{lazy_sample}} objects.}
\item{test}{A list column with objects representing the training sets. For \code{crossv_kfold}, the elements are vectors of the indexes; for \code{crossv_kfold_n}, the elements are \code{\link[=LazySample]{lazy_sample}} objects.}
}
}
\description{
Generate cross-validated K-fold test-training pairs. In addition to ordinary
K-fold cross-validation, this supports stratified K-fold cross validation
if \code{data} is a grouped data frame and \code{stratify = TRUE}, and
Group K-fold if \code{data} is a grouped data frame and \code{stratify = FALSE}.
}
\references{
\itemize{
\item{Breiman, L., Friedman, J.H., Olshen, R.A. and Stone, C.J. (1984) Classification and Regression Trees. Wadsworth.}
\item{Burman, P. (1989) A comparative study of ordinary cross-validation, v-fold cross-validation and repeated learning-testing methods. Biometrika, 76, 503–514}
\item{Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.}
\item{Efron, B. (1986) How biased is the apparent error rate of a prediction rule? Journal of the American Statistical Association, 81, 461–470.}
\item{Stone, M. (1974) Cross-validation choice and assessment of statistical predictions (with Discussion). Journal of the Royal Statistical Society, B, 36, 111–147.}
}
}
\seealso{
This function has more features than the \pkg{modelr} function
  \code{\link[modelr]{crossv_kfold}}.
}
