% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_loo.R
\name{crossv_loo}
\alias{crossv_loo}
\alias{crossv_loo.data.frame}
\alias{crossv_loo.grouped_df}
\alias{crossv_lpo}
\alias{crossv_lpo.data.frame}
\alias{crossv_lpo.grouped_df}
\title{Leave-p-out cross-validation}
\usage{
crossv_loo(data, ...)

\method{crossv_loo}{data.frame}(data, ...)

\method{crossv_loo}{grouped_df}(data, ...)

crossv_lpo(data, p, ...)

\method{crossv_lpo}{data.frame}(data, p = 1, ...)

\method{crossv_lpo}{grouped_df}(data, p = 1, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{p}{Number of observations or groups to leave out}
}
\value{
A data frame with n rows and columns \code{test} and \code{train}.
  \code{test} and \code{train} are list-columns containing
  \code{\link{resample}} objects.
}
\description{
Leave-one-out and leave-p-out cross-validation.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Leave-one-out cross validation

\item \code{grouped_df}: Leave-one-group out cross validation

\item \code{data.frame}: Leave-p-out out cross validation

\item \code{grouped_df}: Leave-p-groups out cross validation
}}

