% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpo.R
\name{crossv_lpo}
\alias{crossv_lpo}
\alias{crossv_lpo.data.frame}
\alias{crossv_lpo.grouped_df}
\alias{crossv_loo}
\title{Generate cross-validated leave-p-out test-training pairs}
\usage{
crossv_lpo(x, p, ...)

\method{crossv_lpo}{data.frame}(x, p = 1L, ...)

\method{crossv_lpo}{grouped_df}(x, p = 1L, ...)

crossv_loo(x)
}
\arguments{
\item{x}{A data frame or vector}

\item{p}{The number of elements to include in the test set.}

\item{...}{Passed to methods}
}
\value{
A data frame with \eqn{\binom{n}{p}}{n choose p} rows and the
following columns:
\describe{
\item{train}{A list of \code{\link{resample}} objects. Training sets.}
\item{test}{A list of \code{\link{resample}} objects. Test sets.}
\item{.id}{Character vector. ID of the fold.}
}
}
\description{
Generate cross-validated leave-p-out (including leave-one-out)
test-training pairs.
}
