% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_mc.R
\name{crossv_mc}
\alias{crossv_mc}
\alias{crossv_mc_n}
\title{Generate K cross-validation test-training pairs}
\usage{
crossv_mc(expr, ..., n = NULL, extractor = NULL)

crossv_mc_n(n, K = 1L, frac = 0.3, size = NULL, prob = NULL)
}
\arguments{
\item{expr}{A quosure or one-sided formula.}

\item{...}{Arguments passed to \code{crossv_mc_n}.}

\item{n}{Integer scalar. The number of elements to sample from. If \code{n} is \code{NULL}, then the number of elements will determined by calling \code{\link{idx_len}} on \code{expr}.}

\item{extractor}{A function or coercible to a function by \code{as_function}. See \code{\link{lazy_sample}}.}

\item{K}{Number of test/train splits to generate.}

\item{frac}{A numeric scalar between 0 and 1. The fraction of elements in the test set.#' @param frac. A numeric scalar between 0 and 1. The fraction of elements in the test set.}

\item{size}{An integer scalar. The number of elements in the test set. If \code{size} is non-\code{NULL}, then it takes preference over \code{frac}.}

\item{prob}{Probability weight that an element is in the \code{test} set.}
}
\description{
\code{crossv_mc} generates random train/test pairs where either proportion \code{frac} of the elements are in the test set and \code{1 - frac} elements are in the training set, or \code{size} elements are in the test set and \code{n - size} elements are in the training set.
}
\references{
\itemize{
\item{Breiman, L., Friedman, J.H., Olshen, R.A. and Stone, C.J. (1984) Classification and Regression Trees. Wadsworth.}
\item{Burman, P. (1989) A comparative study of ordinary cross-validation, v-fold cross-validation and repeated learning-testing methods. Biometrika, 76, 503–514}
\item{Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.}
\item{Efron, B. (1986) How biased is the apparent error rate of a prediction rule? Journal of the American Statistical Association, 81, 461–470.}
\item{Stone, M. (1974) Cross-validation choice and assessment of statistical predictions (with Discussion). Journal of the Royal Statistical Society, B, 36, 111–147.}
}
}
\seealso{
The \pkg{modelr} function \code{\link[modelr]{crossv_mc}}.

Other cross-validation functions: \code{\link{crossv_idx}},
  \code{\link{crossv_lpo}}
}
