% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_mc.R
\name{crossv_mc}
\alias{crossv_mc}
\alias{crossv_mc.data.frame}
\alias{crossv_mc.grouped_df}
\title{Generate cross-validated test-training pairs}
\usage{
crossv_mc(data, n, ...)

\method{crossv_mc}{data.frame}(data, n, test = 0.3, ...)

\method{crossv_mc}{grouped_df}(data, n, test = 0.3, stratify = FALSE, ...)
}
\arguments{
\item{data}{A data frame}

\item{n}{Number of replicates}

\item{...}{Arguments passed to methods}

\item{test}{Proportion of observation in the test sample}

\item{stratify}{If \code{TRUE}, then test-train splits are done
within each group so that the overall test-train split has
approximately equal proportions of each group. If \code{FALSE},
then test-train splits by groups.}
}
\value{
A data frame with \code{n} rows and the following columns:
\describe{
\item{train}{A list of \code{\link{resample}} objects. Training sets.}
\item{test}{A list of \code{\link{resample}} objects. Test sets.}
\item{.id}{Character vector. ID of the fold.}
}
}
\description{
Generate cross-validated test-training pairs
}
