% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_tskfold.R
\name{crossv_tskfold}
\alias{crossv_tskfold}
\title{Generate cross-validated time-series K-fold test-training pairs}
\usage{
crossv_tskfold(data, K, ...)
}
\arguments{
\item{data}{A data frame or vector}

\item{K}{The number of folds}

\item{...}{Passed to methods}
}
\description{
Splits data into test/train sets for cross-validating time series
data. In each split, the indices are increasing, so only previous data
are used to test future data.
}
\details{
The data are split into \eqn{K \geq 2} folds, where for each
\eqn{i \in 2:K}{i in 2:K}, the \eqn{i}-th fold is the test set
and folds \eqn{1, ..., (i - 1)} folds are the training set.
}
\references{
\itemize{
\item{Scikit-learn v. 0.18.1. Cross-validation: evaluating estimator performance.
   \href{http://scikit-learn.org/stable/modules/cross_validation.html#cross-validation}{URL}.}
}
}
