% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskfold.R
\name{crossv_tskfold}
\alias{crossv_tskfold}
\alias{crossv_tskfold.data.frame}
\alias{crossv_tskfold.grouped_df}
\title{Generate cross-validated time-series K-fold test-training pairs}
\usage{
crossv_tskfold(data, K, ...)

\method{crossv_tskfold}{data.frame}(data, K = 5L, ...)

\method{crossv_tskfold}{grouped_df}(data, K = 5L, ...)
}
\arguments{
\item{data}{A data frame or vector}

\item{K}{The number of folds}

\item{...}{Passed to methods}
}
\value{
A data frame \code{K} following columns:
\describe{
\item{train}{A list of \code{\link{resample}} objects. Training sets.}
\item{test}{A list of \code{\link{resample}} objects. Test sets.}
\item{.id}{An integer vector of identifiers.}
}
}
\description{
Splits data into test/train sets for cross-validating time series
data. In each split, the indices are increasing, so only previous data
are used to test future data.
}
\details{
The data are split into \eqn{K \geq 2} folds, where for each
\eqn{i \in 2:K}{i in 2:K}, the \eqn{i}-th fold is the test set
and folds \eqn{1, ..., (i - 1)} folds are the training set.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Rows (observations) are split into folds.
The ordering of the data frame is assmed to be meaningful.

\item \code{grouped_df}: The groups are split into folds. The
ordering of the groups is assumed to be meaningful. No stratification
option is provided since the cases in which stratification can usually
be rewritten to operate on the groups rather than within groups.
}}

