% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskfold.R
\name{crossv_tskfold}
\alias{crossv_tskfold}
\alias{crossv_tskfold.grouped_df}
\title{Generate cross-validated time-series k-fold test-training pairs}
\usage{
crossv_tskfold(x, k, ...)

\method{crossv_tskfold}{grouped_df}(x, k = 5L, ...)
}
\arguments{
\item{x}{A data frame or vector}

\item{k}{The number of folds}

\item{...}{Passed to methods}
}
\value{
A data frame with \code{k} rows and the following columns:
\describe{
\item{train}{A list of \code{\link{resample}} objects. Training sets.}
\item{test}{A list of \code{\link{resample}} objects. Test sets.}
\item{.id}{Character vector. ID of the fold.}
}
}
\description{
Splits data into test/train sets for cross-validating time series
data. In each split, the indices are increasing, so only previous data
are used to test future data.
}
\details{
The data are split into \eqn{k \geq 2} folds, where for each
\eqn{i \in 2:k}{i in 2:k}, the \eqn{i}-th fold is the test set
and the \eqn{1:(i - 1)}-th folds are the training set.
}
\section{Methods (by class)}{
\itemize{
\item \code{grouped_df}: The groups are split into folds. The
ordering of the groups is assumed to be meaningful. No stratification
option is provided since the cases in which stratification can usually
be rewritten to operate on the groups rather than within groups.
}}

