% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\title{Generate jackknife replicates}
\usage{
jackknife(data, id = ".id")
}
\arguments{
\item{data}{A data frame}

\item{id}{Name of variable that gives each model a unique integer id.}
}
\value{
A data frame containing a single column, \code{jackknife},
  which is a list-columns containing \code{\link{resample}}
}
\description{
Generate \code{n} jacknife replicates, each of which is \code{n - 1}
rows.
}
