% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\alias{jackknife.data.frame}
\alias{jackknife.grouped_df}
\title{Generate jackknife replicates}
\usage{
jackknife(data, ...)

\method{jackknife}{data.frame}(data, ...)

\method{jackknife}{grouped_df}(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}
}
\value{
A data frame containing a single column, \code{jackknife},
  which is a list-columns containing \code{\link{resample}}
}
\description{
Generate \code{n} jacknife replicates, each of which is \code{n - 1}
rows. For grouepd data frames, this generates \code{n_groups} replicates,
with each repliacate dropping all rows in a group.
}
