% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\alias{jackknife.data.frame}
\alias{jackknife.grouped_df}
\title{Generate Jackknife Subsamples}
\usage{
jackknife(x, p, ...)

\method{jackknife}{data.frame}(x, p = 1L, ...)

\method{jackknife}{grouped_df}(x, p = 1L, ...)
}
\arguments{
\item{x}{A data frame or vector}

\item{p}{The number of elements to delete; \code{p = 1} is the standard
jackknife, while \code{p > 1} is the delete-p jackknife.}

\item{...}{Passed to methods}
}
\value{
A data frame with \eqn{{n}\choose{p}}{n choose p} rows and the
following columns:
\describe{
\item{sample}{A list of \code{\link{resample}} objects.}
\item{.id}{Character vector. ID of the sample.}
}
}
\description{
Generate jackknife subsamples.
}
