% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\alias{jackknife.default}
\alias{jackknife.grouped_df}
\title{Generate Jackknife Subsamples}
\usage{
jackknife(data, p, ...)

\method{jackknife}{default}(data, p = 1L, ...)

\method{jackknife}{grouped_df}(data, p = 1L, ...)
}
\arguments{
\item{data}{A data frame or vector}

\item{p}{The number of elements to delete. \code{p = 1} is the ordinary
(delete-one) jackknife. \code{p > 1} is the block (delete-\eqn{p})
jackknife.}

\item{...}{Arguments passed to methods}
}
\description{
Generate jackknife subsamples. For data with \eqn{n} elements,
the jackknife generates \eqn{n} subsamples, each with one element
deleted. The generalized (delete-\eqn{p}, block) jackknife generates
subsamples by deleting \code{p} observations.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Delete rows from a data frame.

\item \code{grouped_df}: Delete groups from a grouped data frame.
}}

\references{
\itemize{
\item{Davison, A. C. & Hinkley, D. V. (1997) Bootstrap Methods and Their Applications. Cambridge University Press, Cambridge. ISBN 0-521-57391-2}
\item{Tukey, J. W. (1958). "Bias and confidence in not quite large samples". The Annals of Mathematical Statistics.  \href{https://dx.doi.org/10.1214/aoms/1177706647}{doi:10.1214/aoms/1177706647}.}
\item{Efron, B.; Stein, C. (May 1981). "The Jackknife Estimate of Variance". The Annals of Statistics. \href{https://dx.doi.org/10.1214/aos/1176345462}{doi:10.1214/aos/1176345462}.}
}
}
