% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\alias{permute_n}
\title{Generate permutation subsets}
\usage{
permute(expr, R = 1L, n = NULL, extractor = NULL)

permute_n(n, R = 1L)
}
\arguments{
\item{expr}{An expression for the data to be sampled. This does not need to be a data frame, as \code{\link[=LazySample]{lazy_sample}} objects support or can be extended to support many classes. The unevaluated expression and its environment are captured as a \code{\link[rlang]{quosure}}.}

\item{R}{The number of permutations to generate.}

\item{n}{A non-negative scalar integer representing the number of elements to sample from. If \code{n} is \code{NULL}, then \code{expr} is evaluated and called be \code{\link{idx_len}} to calculate this number. However, if \code{n} is non-\code{NULL}, then \code{expr} is not evaluated at all until samples are requested.}

\item{extractor}{A function or object coercible to a function by \code{\link[rlang]{as_function}}. See \link{LazySample} for more information.}

\item{...}{Arguments passed to methods}
}
\value{
data frame (\code{\link[tibble]{tibble}}) with \code{R} rows and the following columns.
\describe{
\item{sample}{A list column with objects representing the training sets. For \code{permute}, the elements are vectors of the indexes; for \code{permute_n}, the elements are \code{\link[=LazySample]{lazy_sample}} objects.}
}
}
\description{
Generate \code{R} replicates with permutations of the rows or groups of a data frame.
}
