% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\alias{permute.data.frame}
\alias{permute.grouped_df}
\title{Generate permutation subsets}
\usage{
permute(data, ...)

\method{permute}{data.frame}(data, R = 1L, ...)

\method{permute}{grouped_df}(data, R = 1L, groups = TRUE, within = 0, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{R}{The number of permutations to generate.}

\item{stratify}{If \code{TRUE}, the permute observations within each group.
Otherwise permute the groups themselves.}
}
\description{
Generate \code{R} replicates with permutations of the rows or groups of a data frame.
}
\details{
If \code{R} is greater than the total number of permutations,
  only the total number of permutations is returned. Since the total number
  of permutations is \code{factorial(n)}, this will only happen if \code{n}
  is small.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Permute rows of a data frame.

\item \code{grouped_df}: Permutes rows within each group in a grouped data frame
if \code{stratify = TRUE}. Otherwise, permute the groups.
}}

