% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permute}
\alias{permute}
\title{Generate permutation replicates}
\usage{
permute(data, n, id = ".id")
}
\arguments{
\item{data}{A data frame}

\item{n}{Number of replicates to draw}

\item{id}{Name of variable that gives each model a unique integer id.}
}
\value{
A data frame with a single column named \code{perm}.
  This column is a list column of \code{\link[modelr]{resample}} objects.
}
\description{
Generate permutation replicates, which randomly shuffles the row of
the data frame.
}
