% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\alias{as.integer.resample}
\alias{as.character.resample}
\alias{c.resample}
\title{A "lazy" resample object}
\usage{
resample(data, idx, ...)

\method{as.integer}{resample}(x, ...)

\method{as.character}{resample}(x, ...)

\method{c}{resample}(...)
}
\arguments{
\item{data}{An object. This can be any object for which a \code{[} method is defined.}

\item{idx}{A vector of integer indexes. For efficiency, this function does not check whether these indices are valid.}

\item{...}{Other arguments}

\item{x}{An object.}
}
\description{
Often you will resample a dataset many times, as in cross-validation or bootstrapping.
Copying and storing the complete resample each time is both time-consuming and memory intensive.
However, all that is needed for the resample
The \code{resample}
instead stores a "pointer" to the original dataset, and a vector of row
indexes.
}
\section{Methods (by class)}{
\itemize{
\item \code{resample}: Return the indexes as an integer vector.

\item \code{resample}: Return the indexes as a character vector.

\item \code{resample}: Concatenate \code{resample} and index objects into a
single \code{resample} objects. The \code{data} element of the first
object is used as the data for the returned obejct, with the \code{data}
elements of other resample objects neither used nor checked.
The elements of \code{...} must be \code{resample} objects, or
\code{integer}, \code{numberic}, or \code{character} vectors, which are
treated as indexes.
}}

\examples{
r <- resample(mtcars, 1:10)
as.integer(r)
as.data.frame(r)
dim(r)
# check that the data part of r is identical to orig data
identical(r, mtcars)
}
