% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Create lazy resample object(s)}
\usage{
resample(data, idx, check = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{idx}{An integer vector}

\item{check}{If \code{TRUE}, then run test that \code{idx} is consistent
with}
}
\value{
A \code{resample} object, a list with two elements: \code{data}
  and \code{idx}.
}
\description{
Create lazy resample object(s)
}
\examples{
resample(mtcars, 1:10)
}
\seealso{
This function is a slight variant of \code{\link[modelr]{resample}}.
  The function \code{\link{resample_lst}} generates a list of resample objects.
}
