% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{resample_bootstrap}
\alias{resample_bootstrap}
\alias{resample_bootstrap.data_frame}
\alias{resample_bootstrap.grouped_df}
\title{Generate a bootstrap replicate}
\usage{
resample_bootstrap(data, ...)

\method{resample_bootstrap}{data_frame}(data, ...)

\method{resample_bootstrap}{grouped_df}(data, stratify = TRUE,
  groups = FALSE, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Passed to methods}

\item{stratify}{Resample within groups (stratified bootstrap)}

\item{groups}{Resample groups (clustered bootstrap)}
}
\value{
A \code{\link[modelr]{resample}} object.
}
\description{
Generate a bootstrap replicate
}
