% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{resample_bootstrap}
\alias{resample_bootstrap}
\title{Generate a bootstrap}
\usage{
resample_bootstrap(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Passed to methods}

\item{within}{Resample within groups (stratified bootstrap)}

\item{groups}{Resample groups (clustered bootstrap)}
}
\value{
A \code{\link[modelr]{resample}} object.
}
\description{
Generate a bootstrap
}
