% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_df.R
\name{resample_df}
\alias{resample_df}
\alias{resample_df.data.frame}
\alias{resample_df.grouped_df}
\title{Generate subsamples from indices}
\usage{
resample_df(data, ...)

\method{resample_df}{data.frame}(data, samples, ...)

\method{resample_df}{grouped_df}(data, samples, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{samples}{A list of integer vectors, with the indices in each subsample.}
}
\value{
A data frame with rows for each subsample and the following columns:
\describe{
\item{train}{A list of \code{\link[modelr]{resample}} objects. Training sets.}
\item{test}{A list of \code{\link[modelr]{resample}} objects. Test sets.}
\item{.id}{An integer vector of identifiers}
}
}
\description{
Generate subsamples from manually specified indices of observations in each
subsample.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Generate subsamples of rows of data frames.

\item \code{grouped_df}: Generate subsamples of groups of grouped data frames.
}}

