% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout.R
\name{resample_holdout}
\alias{resample_holdout}
\alias{resample_holdout.grouped_df}
\title{Generate test/train resample objects}
\usage{
resample_holdout(data, ...)

\method{resample_holdout}{grouped_df}(data, test = NULL, train = NULL, ...)
}
\arguments{
\item{data}{A data table}

\item{...}{Arguments passed to methods}

\item{test, train}{Integer vectors with the indexes of the test
and training samples.
In \code{resample_holdout.data.frame}, \code{test} and \code{train}
are the indexes of observations.
In \code{resample_holdout.grouped_df}, \code{test} and \code{train}
are group numbers.
One of \code{test} or \code{train} must be non-\code{NULL}.}
}
\value{
A named list of two \code{\link[modelr]{resample}} objects
  for the "test" and "train" sets.
}
\description{
Generate test/train resample objects
}
\section{Methods (by class)}{
\itemize{
\item \code{grouped_df}: Resamples
}}

