% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_idx.R
\name{resample_idx}
\alias{resample_idx}
\title{Generate subsamples from indices}
\usage{
resample_idx(data, idx, ...)
}
\arguments{
\item{data}{An object}

\item{idx}{A list of integer vectors or an integer vector, with the indices in each subsample.}

\item{...}{Arguments passed to methods}
}
\value{
A data frame with rows for each subsample and the following columns:
\describe{
\item{sample}{A list of \code{\link[modelr]{resample}} objects. Training sets.}
\item{.id}{An integer vector of identifiers.}
}
}
\description{
Generate subsamples from manually specified indices of observations in each
subsample.
}
\examples{
resample(mtcars, 1:10)
}
