% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_lst}
\alias{resample_lst}
\title{Create a list of resample objects}
\usage{
resample_lst(data, idx)
}
\arguments{
\item{data}{A data frame}

\item{idx}{A list of integer vectors of indexes.}
}
\value{
A \code{list} of \code{\link[modelr]{resample}} objects.
}
\description{
Create a list of resample objects
}
\examples{
resample_lst(mtcars, list(1:3, 4:6, 7:10))
}
\seealso{
\code{\link{resample}} generates a single resample object.
}
