% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\alias{roll.data.frame}
\alias{roll.grouped_df}
\title{Generate rolling windows}
\usage{
roll(data, ...)

\method{roll}{data.frame}(data, width = 1L, by = 1L, from = 1L,
  to = nrow(data), partial = FALSE, align = c("center", "left", "right"),
  offsets = NULL, ...)

\method{roll}{grouped_df}(data, width = 1L, by = 1L, from = 1L,
  to = dplyr::n_groups(data), partial = FALSE, align = c("center", "left",
  "right"), offsets = NULL, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{width}{An integer specifying the window width.}

\item{by}{Windows are at every by-th point rather than every observation}

\item{partial}{If \code{FALSE}, then windows which extend outside the range
of the data are not included. If \code{TRUE}, then those windows outside
the range of the data are included, though they are truncated.}

\item{align}{Is the window aligned to the left, right, or center.}
}
\value{
A data frame with four columns
\describe{
\item{window}{A list column of \code{\link[modelr]{resample}} objects}
\item{.id}{integer: the index position of the window}
\item{.width}{integer: the width of the window}
\item{.align}{character: alignment of the window. "l" for left,
"c" for center, "r" for right}
}
}
\description{
Generate \code{\link[modelr]{resample}} objects for rolling windows.
For data frames, the windows are applied row-wise, for grouped data frames,
the windows are applied group-wise.
}
\seealso{
The \code{\link[zoo]{rollapply}} function in \pkg{zoo}
}
