% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\alias{roll.data.frame}
\alias{roll.grouped_df}
\title{Generate rolling windows}
\usage{
roll(data, ...)

\method{roll}{data.frame}(data, width = 1L, align = c("left", "right",
  "center"), partial = TRUE, indices = NULL, from = 1L, to = n,
  by = 1L, offsets = NULL, ...)

\method{roll}{grouped_df}(data, width = 1L, align = c("left", "right",
  "center"), partial = TRUE, indices = NULL, from = 1L, to = n,
  by = 1L, offsets = NULL, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{width}{The window width}

\item{align}{Is the window left-aligned, right-aligned, or centered?}

\item{partial}{logical or numeric. If \code{FALSE} all indexes of a rolling
window must be within the rows (or groups) of \code{data}. If \code{TRUE},
then the subset of indexes that are in range are used.
A numeric argument to \code{partial} can be used to determine the minimal
window size for partial windows.}

\item{indices, from, to, by}{The indices at which to generate windows. If
\code{indices} is non-\code{NULL}, then it is used. Otherwise, the indices
are generated from \code{seq(from, to, by)}.}

\item{offsets}{numeric. If not-\code{NULL}, then the offsets of the window
indices relative to the reference index. If \code{NULL}, then the window
offsets are generated from the \code{width} and \code{align} arguments.}
}
\value{
A data frame
}
\description{
Generate resample objects of rolling windows over elements of a data frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Data frame method. This rows are assumed to be ordered.

\item \code{grouped_df}: Grouped data frame method. The groups are assumed to be ordered, and
the cross validation works on groups rather than rows.
}}

