% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbootstrap.R
\name{tsbootstrap}
\alias{tsbootstrap}
\alias{tsbootstrap.data.frame}
\alias{tsbootstrap.grouped_df}
\title{Generate time-series bootstraps replicates}
\usage{
tsbootstrap(data, ...)

\method{tsbootstrap}{data.frame}(data, k = 1L, size = 1L, ...)

\method{tsbootstrap}{grouped_df}(data, k = 1L, size = 1L, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Arguments passed to methods}

\item{k}{Number of bootstrap replicates}

\item{size}{Block length size}
}
\description{
Generate time-series bootstraps replicates
}
\details{
Bergmeir et al. (2016) describe a moving block bootstrap (MBB) method.
Suppose a time series of size \eqn{n} and a chosen block size of size \eqn{l}.
Randomly sample \eqn{\lfloor n / l \rfloor + 2}{floor(n / l) + 2} blocks of
size \eqn{l} and concatenate them into a single series
Then randomly discard 0 to \eqn{l - 1} values from the start of the series,
and truncate the series to length \eqn{n}. The later correction ensures that
every value from the original series can be placed anywhere in the bootstrapped
series.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Method for data frames, sampling the rows of the data frame.

\item \code{grouped_df}: Method for grouped data frames.
This samples groups instead of rows.
}}

\references{
\itemize{
\item{Kunsch, H.R. (1989) The jackknife and the bootstrap for general stationary observations. \emph{Annals of Statistics}}
\item{Lahiri, S., 2003. \emph{Resampling Methods for Dependent Data}. Springer}
\item{Bergmeir, Christoph, Rob J. Hyndman, and Jose M. Benitz. 2016. "Bagging Exponential Smoothing Methods using STL Decomposition and Box-Cox Transformation." \emph{International Journal of Forecasting} \href{http://dx.doi.org/10.1016/j.ijforecast.2015.07.002}{doi:10.1016/j.ijforecast.2015.07.002}.}
}
}
